Welp just confused myself with my own method name so let's not
present me: "derrp what is this occurrence_ident keeping track of, i'm
just about to add keeping track of the idents that have been used, that
isn't here yet wtf"

past me: "oh hey i need to get an identifier of the thing that just
occurred, so let's call this concept in the metaverse 'ident' that
totally won't get conflated in my brain with the concept of a rust
ident!"

present me: "past me, you are a poop"

past me can't respond because i don't have a time machine. take that,
past me!